requirements.txt
django==3.0.2
djangorestframework==3.11.0
django-extra-fields==2.0.2